extract primary ticket identifying into own service
