+++ app/services/zendesk/duplicate_intake_matcher.rb
@@ -41,7 +41,7 @@ module Zendesk
       headers = ["name", "email", "phone", "filing years", "intake ticket mapping", "primary ticket id"]
       with_primaries = strong_matches.lazy.map do |matches|
         mapping = matches.intake_ids.zip(matches.ticket_ids).to_h
-        primary = merging_service.find_primary_ticket(matches.ticket_ids.compact)&.id
+        primary = identifying_service.find_primary_ticket(matches.ticket_ids.compact)&.id
         [
           matches.lower_name,
           matches.lower_email,
@@ -69,6 +69,10 @@ module Zendesk
 
     def merging_service
       @merging_service ||= TicketMergingService.new
+    end
+
+    def identifying_service
+      @identifying_service ||= TicketIdentifyingService.new
     end
   end
 end
